feat: add buy sell order action on home
