+++ web-registry/src/features/marketplace/BuySellOrderFlow.tsx
@@ -1,0 +1,130 @@
+import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { DeliverTxResponse } from '@cosmjs/stargate';
+
+import { CelebrateIcon } from 'web-components/lib/components/icons/CelebrateIcon';
+import { ProcessingModal } from 'web-components/lib/components/modal/ProcessingModal';
+import { TxErrorModal } from 'web-components/lib/components/modal/TxErrorModal';
+import { Item } from 'web-components/lib/components/modal/TxModal';
+import { TxSuccessfulModal } from 'web-components/lib/components/modal/TxSuccessfulModal';
+
+import { getHashUrl } from 'lib/block-explorer';
+
+import useBuySellOrderSubmit from 'pages/Marketplace/Storefront/hooks/useBuySellOrderSubmit';
+import { VIEW_ECOCREDITS } from 'pages/Projects/Projects.config';
+import { ProjectWithOrderData } from 'pages/Projects/Projects.types';
+import { Link } from 'components/atoms';
+import { BuyCreditsModal } from 'components/organisms';
+import { useMsgClient } from 'hooks';
+
+type Props = {
+  selectedProject: ProjectWithOrderData | null;
+};
+
+export const BuySellOrderFlow = ({ selectedProject }: Props): JSX.Element => {
+  const [isBuyModalOpen, setIsBuyModalOpen] = useState(false);
+  const [isProcessingModalOpen, setIsProcessingModalOpen] = useState(false);
+  const [txModalTitle, setTxModalTitle] = useState<string>('');
+  const [txButtonTitle, setTxButtonTitle] = useState<string>('');
+  const [txModalHeader, setTxModalHeader] = useState<string>('');
+  const [cardItems, setCardItems] = useState<Item[] | undefined>(undefined);
+  const navigate = useNavigate();
+
+  const closeBuyModal = (): void => setIsBuyModalOpen(false);
+  const closeProcessingModal = (): void => setIsProcessingModalOpen(false);
+
+  const handleTxQueued = (): void => {
+    setIsProcessingModalOpen(true);
+  };
+  const handleTxModalClose = (): void => {
+    setCardItems(undefined);
+    setTxModalTitle('');
+    setTxModalHeader('');
+    setDeliverTxResponse(undefined);
+    setError(undefined);
+  };
+  const handleError = (): void => {
+    closeProcessingModal();
+    setTxModalTitle('Buy Credits Error');
+  };
+  const handleTxDelivered = async (
+    _deliverTxResponse: DeliverTxResponse,
+  ): Promise<void> => {
+    closeProcessingModal();
+    closeBuyModal();
+  };
+  const onTxSuccessButtonClick = (): void => {
+    handleTxModalClose();
+    navigate('/ecocredits/dashboard');
+  };
+
+  useEffect(() => {
+    if (selectedProject) {
+      setIsBuyModalOpen(true);
+    }
+  }, [selectedProject]);
+
+  const {
+    signAndBroadcast,
+    setDeliverTxResponse,
+    wallet,
+    deliverTxResponse,
+    error,
+    setError,
+  } = useMsgClient(handleTxQueued, handleTxDelivered, handleError);
+  const accountAddress = wallet?.address;
+  const txHash = deliverTxResponse?.transactionHash;
+  const txHashUrl = getHashUrl(txHash);
+
+  const buySellOrderSubmit = useBuySellOrderSubmit({
+    accountAddress,
+    signAndBroadcast,
+    setCardItems,
+    setTxButtonTitle,
+    setTxModalHeader,
+    setTxModalTitle,
+    buttonTitle: VIEW_ECOCREDITS,
+  });
+
+  return (
+    <>
+      <BuyCreditsModal
+        open={isBuyModalOpen}
+        onClose={closeBuyModal}
+        onSubmit={buySellOrderSubmit}
+        project={{
+          id: selectedProject?.id.toString() ?? '',
+          sellOrders: selectedProject?.sellOrders,
+        }}
+      />
+      <ProcessingModal
+        open={isProcessingModalOpen}
+        onClose={closeProcessingModal}
+      />
+      <TxSuccessfulModal
+        open={!!txHash && !error}
+        onClose={handleTxModalClose}
+        txHash={txHash ?? ''}
+        txHashUrl={txHashUrl}
+        title={txModalHeader}
+        cardTitle={txModalTitle}
+        buttonTitle={txButtonTitle}
+        cardItems={cardItems}
+        linkComponent={Link}
+        onButtonClick={onTxSuccessButtonClick}
+        icon={<CelebrateIcon sx={{ width: '85px', height: '106px' }} />}
+      />
+      <TxErrorModal
+        error={error ?? ''}
+        open={!!error && (!!txModalTitle || !!deliverTxResponse)}
+        onClose={handleTxModalClose}
+        txHash={txHash ?? ''}
+        txHashUrl={txHashUrl}
+        cardTitle={txModalTitle}
+        linkComponent={Link}
+        onButtonClick={handleTxModalClose}
+        buttonTitle="close"
+      />
+    </>
+  );
+};
